+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 # Wikibase DataModel release notes
 
-## Version 1.0 (dev)
+## Version 1.0 (2014-09-02)
 
 #### Breaking changes